Refactor.

+++ monoruby/src/executor/compiler/jitgen/method_call.rs
@@ -536,8 +536,8 @@ impl Codegen {
         }
         self.jit_handle_error(ctx, pc);
         monoasm! { &mut self.jit,
-            lea  rdx, [rax + ((RVALUE_OFFSET_KIND + PROCINNER_FUNCDATA))];
-            movq rax, [rax + ((RVALUE_OFFSET_KIND + PROCINNER_OUTER))];
+            lea  rdx, [rax + ((RVALUE_OFFSET_KIND as i64 + PROCINNER_FUNCDATA))];
+            movq rax, [rax + ((RVALUE_OFFSET_KIND as i64 + PROCINNER_OUTER))];
             // rax <- outer_cfp, rdx <- &FuncData
         }
